src/utils/requirements.txt
requests
python-dotenvrequests
python-dotenv
